% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_download.R
\name{pv_download}
\alias{pv_download}
\title{Download PatentsView Data Files}
\usage{
pv_download(url = NULL, dest = NULL, timeout = 600)
}
\arguments{
\item{url}{url of the file to download}

\item{dest}{destination for the files}

\item{timeout}{time window for download}
}
\value{

}
\description{
Download one or more PatentsView raw files using urls from a meta data table
created by \code{pv_meta()}
}
\examples{
\dontrun{pv_download(grant_meta$url[1], dest = "grants", timeout = 600)}
}
