% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_import.R
\name{pv_import}
\alias{pv_import}
\title{Import PatentsView Data Tables}
\usage{
pv_import(path = NULL, meta_path = NULL, save_as = NULL, dest = NULL)
}
\arguments{
\item{path}{path to file to import}

\item{meta_path}{path to metadata file to validate nrows after import}

\item{save_as}{file type to save as: "rds", "rda", "qs", "csv", "parquet"}

\item{dest}{destination for save_as, defaults to working directory}
}
\value{
data.frame
}
\description{
Import a PatentsView data table from \code{pv_download()}.
}
\details{
The function uses the vroom package to import tab separated files
downloaded from the PatentsView data download site.
}
\examples{
\dontrun{applications <- pv_import("grants/application.tsv.zip",
meta_path = "grants/meta_grants.rds", save_as = ".qs", dest = "grants")}
}
